backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
boto3==1.40.66
botocore==1.40.66
click==8.3.0
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.121.0
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
httptools==0.7.1
idna==3.11
iniconfig==2.3.0
jmespath==1.0.1
packaging==25.0
pluggy==1.6.0
protobuf==6.33.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
